requirements.txt
google-adk==1.7.0
requests
